Run both certificate.d script and tls-profile script
On cert change, run both certificate.d script and tls-profile script for
all services found for CERT_ID.

Previously. If the service was associated with a tls-profile, only that
script was executed and not the certificate.d script.

Add missing DSM7 certificate.d paths. This is the cause of reports of
services not beeing reloaded after cert updates.

Fixes #20