Maintain cert_id in INFO file on each run.
It was reported in #11 that in some circumstances the cert_id reference
could be missing in /usr/syno/etc/certificate/_archive/INFO.
Before this change the INFO file was only maintained on first run.
Now the the INFO file will be checked on each run and updated if the
cert_id is missing.